Add files via upload
Amazon S3 is used for storing the raw data. The data is then transformed using AWS Glue and Amazon Athena. Finally, the transformed data is loaded into an Amazon Redshift data warehouse for analytics and querying. The project's setup instructions outline a clear process for a user to get the system running, including cloning the repository, setting up a Python virtual environment, installing dependencies, and using